python_backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
numpy==1.21.0
opencv-python==4.5.3.56
mediapipe==0.8.9.1
tensorflow==2.6.0
pillow==8.3.1
scikit-learn==0.24.2
matplotlib==3.4.3
